+++ packages/shopify/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @weaverse/shopify
 
+## 5.3.0
+
+### Minor Changes
+
+- update schema validation
+
+### Patch Changes
+
+- Updated dependencies
+  - @weaverse/react@5.3.0
+
+## 5.2.5
+
+### Patch Changes
+
+- @weaverse/react@5.2.5
+
 ## 5.2.4
 
 ### Patch Changes